+++ PROJECT.py
@@ -33,14 +33,22 @@ class PROJECT(PROJECT_BASE):
     FIXME: list[str] = [
     ]
     # -----------------------------------------------------------------------------------------------------------------
-    VERSION = Version((0, 2, 15))
+    VERSION = Version((0, 2, 16))
     NEWS: list[str] = [
-        "[TESTS] move all near source code",
-        ["[AnnotAttrAux] BIG REF",
-            "make nested Annots from Attrs +use same meth in both",
-            "rename sai_*/gai_",
-            "add values__reinit_mutable +NestInit_Mutable",
+        "[TextFormated] create",
+        "[ReAttempts] create",
+        ["[Attrs+Annots]",
+            "combine two classes finally in one",
+            "add annots_ensure +annots__append",
+            "separate NestRepr__ClsName_SelfStr",
         ],
+        "[IterAux] add get_first_is_not_none",
+        "[DateTime] add UPDATE_ON_STR+DEF_STR_FORMAT",
+
+        ["[Nest*]",
+            "add NestCall_Other",
+            "add NestStR_AttrsPattern",
+         ],
     ]